alterando valores de variaveis para int

+++ src/com/ian/Game.java
@@ -94,7 +94,7 @@ public class Game extends Canvas implements Runnable, KeyListener, MouseListener
             if (this.saveGame) {
                 this.saveGame = false;
                 String[] opt1 = {"level", "vida", "x", "y"};
-                double[] opt2 = {this.CUR_LEVEL, player.getLife(), player.getX(), player.getY()};
+                int[] opt2 = {this.CUR_LEVEL, (int)player.getLife(), player.getX(), player.getY()};
                 GameMenu.saveGame(opt1, opt2, 10);
                 System.out.println("Jogo salvo com sucesso!");
             }
